Stripped down to standalone Python package.
