+++ .vscode/settings.json
@@ -1,6 +1,5 @@
 {
     "python.analysis.extraPaths": [
-        "${workspaceFolder}/sense_aruco/src/",
-        "/opt/ros/noetic/lib/python3/dist-packages"
+        "${workspaceFolder}/src/"
     ]
 }
