--- aruco_marker_description/CMakeLists.txt
@@ -1,9 +1,0 @@
-cmake_minimum_required(VERSION 3.0.0)
-project(aruco_marker_description)
-
-find_package(catkin REQUIRED)
-
-catkin_package()
-
-install(DIRECTORY models
-  DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
