--- aruco_marker_description/models/aruco_fam4x4_1000_id25_sdf/materials/scripts/aruco_fam4x4_1000_id25_sdf.material
@@ -1,17 +1,0 @@
-material aruco_fam4x4_1000_id25_sdf/Image
-{
-  technique
-  {
-    pass
-    {
-      ambient 0.5 0.5 0.5 1.0
-      diffuse 1.0 1.0 1.0 1.0
-      specular 0.0 0.0 0.0 1.0 0.5
-      texture_unit
-      {
-        texture aruco_fam4x4_1000_id25_sdf.jpg
-        filtering trilinear
-      }
-    }
-  }
-}
