--- aruco_marker_description/models/aruco_fam4x4_1000_id25_sdf/model.config
@@ -1,8 +1,0 @@
-<model>
-    <name>aruco_fam4x4_1000_id25_sdf</name>
-    <version>1.0</version>
-    <sdf version="1.4">model.sdf</sdf>
-    <description>
-        Aruco marker ID 25 from family 4x4_1000 as an SDF model.
-    </description>
-</model>
