--- aruco_marker_description/models/aruco_fam4x4_1000_id25_sdf/model.sdf
@@ -1,43 +1,0 @@
-<sdf version="1.6">
-  <model name="aruco_fam4x4_1000_id25_sdf">
-    <pose>0 0 0 0 0 0</pose>
-    <static>true</static>
-    <link name="link">
-      <collision name="collision">
-        <geometry>
-          <plane>
-            <normal>0 0 1</normal>
-            <size>0.1 0.1</size>
-          </plane>
-        </geometry>
-        <surface>
-          <friction>
-            <ode>
-              <mu>100</mu>
-              <mu2>50</mu2>
-            </ode>
-          </friction>
-        </surface>
-      </collision>
-      <visual name="visual">
-        <cast_shadows>false</cast_shadows>
-        <geometry>
-          <plane>
-            <normal>0 0 1</normal>
-            <size>0.36 0.36</size>
-          </plane>
-        <mesh>
-          <uri>model://aruco_fam4x4_1000_id25_sdf/materials/textures/aruco_fam4x4_1000_id25_sdf.dae</uri>
-        </mesh>
-        </geometry>
-        <material>
-          <script>
-            <uri>model://aruco_fam4x4_1000_id25_sdf/materials/scripts/</uri>
-            <uri>model://aruco_fam4x4_1000_id25_sdf/materials/textures/</uri>
-            <name>aruco_fam4x4_1000_id25_sdf/Image</name>
-          </script>
-        </material>
-      </visual>
-    </link>
-  </model>
-</sdf>
