--- aruco_marker_description/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-    <name>aruco_marker_description</name>
-    <version>0.1.0</version>
-    <description>A collection of ArUco markers to use in Gazebo.</description>
-    <maintainer email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</maintainer>
-    <maintainer email="willkode@outlook.com">Troi Williams</maintainer>
-    <license>MIT</license>
-
-    <url type="website">https://github.com/troiwill/sensing-aruco</url>
-    <url type="bugtracker">https://github.com/troiwill/sensing-aruco/issues</url>
-
-    <author email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</author>
-    <author email="willkode@outlook.com">Troi Williams</author>
-
-    <buildtool_depend>catkin</buildtool_depend>
-    <exec_depend>gazebo_ros</exec_depend>
-
-    <export>
-        <gazebo_ros
-            gazebo_media_path="${prefix}"
-            gazebo_model_path="${prefix}/models"
-            /> 
-    </export>
-</package>
