+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "sense_aruco"
-version = "0.0.1"
+version = "0.1.0"
 authors = [
   { name="Troi Williams", email="willkode@outlook.com" },
 ]
