--- sense_aruco/CMakeLists.txt
@@ -1,9 +1,0 @@
-cmake_minimum_required(VERSION 3.0.0)
-project(sense_aruco)
-
-find_package(catkin REQUIRED
-)
-
-catkin_python_setup()
-
-catkin_package()
