--- sense_aruco/package.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>sense_aruco</name>
-  <version>0.1.0</version>
-  <description>Python code for detecting Aruco markers.</description>
-  <maintainer email="willkode@outlook.com">Troi Williams</maintainer>
-  <maintainer email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</maintainer>
-  <license>MIT</license>
-
-  <url type="website">https://github.com/troiwill/sensing-aruco</url>
-  <url type="bugtracker">https://github.com/troiwill/sensing-aruco/issues</url>
-
-  <author email="willkode@outlook.com">Troi Williams</author>
-  <author email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-</package>
