--- sense_aruco/setup.py
@@ -1,9 +1,0 @@
-from distutils.core import setup
-from catkin_pkg.python_setup import generate_distutils_setup
-
-d = generate_distutils_setup(
-    packages=['sense_aruco'],
-    package_dir={'': 'src'}
-)
-
-setup(**d)
