--- sense_aruco_ros/CMakeLists.txt
@@ -1,34 +1,0 @@
-cmake_minimum_required(VERSION 3.0.0)
-project(sense_aruco_ros)
-
-find_package(catkin REQUIRED
-    message_generation
-    geometry_msgs
-    sensor_msgs
-    std_msgs
-)
-
-add_message_files(
-    FILES
-    Marker.msg
-    MarkerArray.msg
-    MarkerPose.msg
-    MarkerPoseArray.msg
-)
-
-generate_messages(
-    DEPENDENCIES
-    std_msgs
-    geometry_msgs
-)
-
-catkin_package(
-    CATKIN_DEPENDS message_runtime
-)
-
-catkin_install_python(PROGRAMS
-    nodes/aruco_detector_node
-    DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
-
-install(DIRECTORY nodes
-    DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
