--- sense_aruco_ros/launch/detect_markers.launch
@@ -1,16 +1,0 @@
-<launch>
-    <arg name="parampath" default="$(find fpv_cam_description)/calib/fpv_cam.yaml"/>
-    <arg name="familyname" default="DICT_4X4_1000"/>
-    <arg name="marker_len" default="0.18"/>
-    <arg name="image_topic" default="/fpv_cam/usb_cam/image_raw"/>
-    <arg name="pose_topic" default="aruco_marker"/>
-
-    <!-- Launch the detector node -->
-    <node name="aruco_detector_node" pkg="sense_aruco_ros" type="aruco_detector_node">
-        <param name="parampath" value="$(arg parampath)"/>
-        <param name="familyname" value="$(arg familyname)"/>
-        <param name="marker_len" value="$(arg marker_len)"/>
-        <param name="image_topic" value="$(arg image_topic)"/>
-        <param name="pose_topic" value="$(arg pose_topic)"/>
-    </node>
-</launch>
