--- sense_aruco_ros/msg/Marker.msg
@@ -1,2 +1,0 @@
-uint32 mid
-float64[8] corners
