--- sense_aruco_ros/msg/MarkerArray.msg
@@ -1,2 +1,0 @@
-Header header
-sense_aruco_ros/Marker[] detections
