--- sense_aruco_ros/msg/MarkerPoseArray.msg
@@ -1,2 +1,0 @@
-Header header
-sense_aruco_ros/MarkerPose[] markers
