--- sense_aruco_ros/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>sense_aruco_ros</name>
-  <version>0.1.0</version>
-  <description>ROS package with messages for detecting Aruco markers.</description>
-  <maintainer email="willkode@outlook.com">Troi Williams</maintainer>
-  <maintainer email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</maintainer>
-  <license>MIT</license>
-
-  <url type="website">https://github.com/troiwill/sensing-aruco</url>
-  <url type="bugtracker">https://github.com/troiwill/sensing-aruco/issues</url>
-
-  <author email="willkode@outlook.com">Troi Williams</author>
-  <author email="sbhogavi@terpmail.umd.edu">Sparsh Bhogavilli</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>message_generation</build_depend>
-  <exec_depend>message_runtime</exec_depend>
-  <exec_depend>cv_bridge</exec_depend>
-  <exec_depend>geometry_msgs</exec_depend>
-  <exec_depend>sensor_msgs</exec_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>sense_aruco</exec_depend>
-
-</package>
