--- sense_aruco_simulations/launch/gazebo.launch
@@ -1,21 +1,0 @@
-<launch>
-    <!-- Gabezo configs -->
-    <arg name="paused" default="false"/>
-    <arg name="use_sim_time" default="true"/>
-    <arg name="gui" default="true"/>
-    <arg name="headless" default="false"/>
-    <arg name="debug" default="false"/>
-    <arg name="verbose" default="false"/>
-    <arg name="respawn_gazebo" default="false"/>
-  
-    <!-- Launch Gazebo world -->
-    <include file="$(find gazebo_ros)/launch/empty_world.launch">
-        <arg name="debug" value="$(arg debug)" />
-        <arg name="gui" value="$(arg gui)" />
-        <arg name="paused" value="$(arg paused)"/>
-        <arg name="use_sim_time" value="$(arg use_sim_time)"/>
-        <arg name="headless" value="$(arg headless)"/>
-        <arg name="verbose" value="$(arg verbose)"/>
-        <arg name="respawn_gazebo" value="$(arg respawn_gazebo)"/>
-    </include>
-</launch>
