--- sense_aruco_simulations/launch/spawn_urdf_xarco.launch
@@ -1,22 +1,0 @@
-<launch>
-    <arg name="model_name" default=""/>
-    <arg name="model_path" default=""/>
-    <arg name="x" default="0"/>
-    <arg name="y" default="0"/>
-    <arg name="z" default="0"/>
-    <arg name="R" default="0"/>
-    <arg name="P" default="0"/>
-    <arg name="Y" default="0"/>
-
-    <!-- <arg name="model" default="$(find realsense2_description)/urdf/test_d435_camera.urdf.xacro"/> -->
-
-    <!-- robot description to load in rviz or gazebo -->
-    <param name="robot_description" command="$(find xacro)/xacro $(arg model_path)" />
-
-    <!-- Spawn marker -->
-    <node name="$(arg model_name)" output="screen" pkg="gazebo_ros"
-        type="spawn_model"
-        args="-urdf -param robot_description -model $(arg model_name) 
-        -x $(arg x) -y $(arg y) -z $(arg z) 
-        -R $(arg R) -P $(arg P) -Y $(arg Y)"/>
-</launch>
