--- sense_aruco_simulations/models/aruco_on_box/model.config
@@ -1,8 +1,0 @@
-<model>
-  <name>aruco_on_box</name>
-  <version>1.0</version>
-  <sdf version='1.6'>model.sdf</sdf>
-  <description>
-    An Aruco marker attached to a box.
-  </description>
-</model>
