--- sense_aruco_simulations/models/box_for_fpv/model.config
@@ -1,8 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-    <name>box_for_fpv</name>
-    <version>1.0</version>
-    <sdf version="1.6">model.sdf</sdf>
-
-    <description>Box for the FPV camera model to rest atop.</description>
-</model>
