--- sense_aruco_simulations/models/box_for_fpv/model.sdf
@@ -1,26 +1,0 @@
-<sdf version="1.6">
-
-    <model name="box_for_fpv">
-        <static>true</static>
-        <pose>0 0 0.24 0 0 0</pose>
-        <link name="link">
-          <collision name="collision">
-            <geometry>
-              <box>
-                <size>0.08 0.08 0.48</size>
-              </box>
-            </geometry>
-          </collision>
-          <visual name="visual">
-            <geometry>
-              <box>
-                <size>0.08 0.08 0.48</size>
-              </box>
-            </geometry>
-            <material>
-              <script>Gazebo/WoodPallet</script>
-            </material>
-          </visual>
-        </link>
-  </model>
-</sdf>
