--- sense_aruco_simulations/models/fpv_cam/model.config
@@ -1,12 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-    <name>fpv_cam</name>
-    <version>1.0</version>
-    <sdf version="1.6">model.sdf</sdf>
-
-    <author>
-        <name>TODO</name>
-        <email>TODO</email>
-    </author>
-    <description>An FPV camera model.</description>
-</model>
